requirements.txt
websockets==11.0.3
python-dotenv==1.0.0
google-cloud-speech==2.24.1
google-cloud-texttospeech==2.15.0
pydub==0.25.1
numpy>=1.26.0
soundfile==0.12.1
google-generativeai==0.3.2
torch>=2.0.0
torchaudio>=2.0.0 